refactor: rename to correct name
