+++ src/record/delete/usecases/__tests__/deleteByRecordNumber/index.test.ts
