+++ src/record/delete/usecases/deleteByRecordNumber/error.ts
