+++ src/record/delete/usecases/deleteByRecordNumber/record.ts
